210725 nej add google login
